requirements.txt
# jupyter==1.1.1
mypy==1.15.0
nb-black-formatter==1.0.1
numpy==2.2.4
pandas==2.2.3
polars==1.26.0
pyarrow==19.0.1
pydantic-settings==2.8.1
ruff==0.11.4
scikit-learn==1.6.1
tqdm==4.67.1
